add classes separation logic
