--- school.rb
@@ -1,47 +1,0 @@
-class Person
-  attr_accessor :name, :age
-  attr_reader :id
-
-  def initialize(age, name = 'Unknown', parent_permission: true)
-    @id = rand(1..100)
-    @name = name
-    @age = age
-    @parent_permission = parent_permission
-  end
-
-  def can_use_services?
-    of_age? || parent_permission == true
-  end
-
-  private
-
-  def of_age?
-    @age >= 18
-  end
-end
-
-class Student < Person
-  attr_accessor :classroom
-
-  def initialize(classroom, age, name = 'Unknown', parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
-    @classroom = classroom
-  end
-
-  def play_hooky
-    '¯\\(ツ)/¯'
-  end
-end
-
-class Teacher < Person
-  attr_accessor :specialization
-
-  def initialize(specialization, age, name = 'Unknown', parent_permission: true)
-    super(age, name, parent_permission: parent_permission)
-    @specialization = specialization
-  end
-
-  def can_use_services?
-    true
-  end
-end
