+++ student.rb
@@ -1,0 +1,13 @@
+require_relative 'person'
+class Student < Person
+  attr_accessor :classroom
+
+  def initialize(classroom, age, name = 'Unknown', parent_permission: true)
+    super(age, name, parent_permission: parent_permission)
+    @classroom = classroom
+  end
+
+  def play_hooky
+    '¯\\(ツ)/¯'
+  end
+end
